backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
django-cors-headers==4.9.0
django-ratelimit==4.1.0
phonenumbers==9.0.18
bcrypt==4.2.0
requests==2.32.5
